+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import "./css/App.css";
 import "./css/common.css";
 import "./css/fontsize.css";
@@ -26,6 +26,10 @@ import { Box } from "@mui/material";
 import { useWidth } from "./components/ContextProvider";
 import { Notification } from "./lib/notification";
 import NotFound from "./pages/notfound";
+import axios from "axios";
+import Verify from "./pages/verify";
+import Resend from "./pages/resend";
+import Recall from "./pages/recall";
 function Source() {
   window.location.replace("https://gitlab.com/metahkg/metahkg");
   return <div />;
@@ -34,16 +38,31 @@ function Telegram() {
   window.location.replace("https://t.me/+WbB7PyRovUY1ZDFl");
   return <div />;
 }
-/*
+/**
  * Menu is not in the Routes to prevent unnecessary rerenders
  * Instead it is controlled by components inside Routes
  */
 export default function App() {
   const [menu] = useMenu();
   const [width] = useWidth();
+  useEffect(() => {
+    if (localStorage.user || localStorage.id) {
+      axios.get("/api/loggedin").then((res) => {
+        if (!res.data.loggedin) {
+          localStorage.removeItem("user");
+          localStorage.removeItem("id");
+          return;
+        }
+        localStorage.user !== res.data.user &&
+          localStorage.setItem("user", res.data.user);
+        localStorage.id !== Number(res.data.id) &&
+          localStorage.setItem("id", res.data.id);
+      });
+    }
+  }, []);
   return (
     <Theme
-      primary={{ main: "#222222" }}
+      primary={{ main: "#222" }}
       secondary={{ main: "#f5bd1f", dark: "#ffc100" }}
     >
       <Notification />
@@ -63,6 +82,8 @@ export default function App() {
               <Route path="/comment/:id" element={<AddComment />} />
               <Route path="/category/:category" element={<Category />} />
               <Route path="/register" element={<Register />} />
+              <Route path="/verify" element={<Verify />} />
+              <Route path="/resend" element={<Resend />} />
               <Route path="/signin" element={<Signin />} />
               <Route path="/create" element={<Create />} />
               <Route path="/search" element={<Search />} />
@@ -71,6 +92,7 @@ export default function App() {
               <Route path="/telegram" element={<Telegram />} />
               <Route path="/profile/:id" element={<Profile />} />
               <Route path="/history/:id" element={<History />} />
+              <Route path="/recall" element={<Recall />} />
               <Route path="/404" element={<NotFound />} />
               <Route path="*" element={<Navigate to="/404" replace />} />
             </Routes>