+++ src/components/conversation/more.tsx
@@ -92,6 +92,7 @@ export default function MoreList(props: {
         </Tooltip>
         <Popper
           open={open}
+          style={{ zIndex: 2147483647 }}
           anchorEl={anchorRef.current}
           role={undefined}
           placement="bottom-start"