+++ src/components/menu.tsx
@@ -12,8 +12,9 @@ import {
   useMenu,
   useSelected,
   useData,
+  useRecall,
 } from "./MenuProvider";
-import { summary } from "../lib/common";
+import { splitarray, summary } from "../lib/common";
 import MenuPreload from "./menu/preload";
 import queryString from "query-string";
 import { useHistory, useNotification, useQuery } from "./ContextProvider";
@@ -21,7 +22,6 @@ import SearchBar from "./searchbar";
 import { useNavigate } from "react-router";
 /**
  * This function renders the main content of the menu
- * @param {AxiosError} err - AxiosError
  */
 function MainContent() {
   const querystring = queryString.parse(window.location.search);
@@ -30,6 +30,7 @@ function MainContent() {
   const [search] = useSearch();
   const [profile] = useProfile();
   const [selected] = useSelected();
+  const [recall] = useRecall();
   const [query] = useQuery();
   const [, setNotification] = useNotification();
   const [id] = useId();
@@ -39,6 +40,7 @@ function MainContent() {
   const [updating, setUpdating] = useState(false);
   const q = decodeURIComponent(String(querystring.q || query || ""));
   const c: string | number = category || `bytid${id}`;
+  const history: any[] = JSON.parse(localStorage.history || "[]") || [];
   /**
    * It sets the notification state to an object with the open property set to true and the text
    * property set to the error message.
@@ -51,14 +53,22 @@ function MainContent() {
     });
     err?.response?.status === 404 && navigate("/404", { replace: true });
   }
+  const mode =
+    (search && "search") ||
+    (profile && "profile") ||
+    (recall && "recall") ||
+    "menu";
   /* A way to make sure that the effect is only run once. */
   useEffect(() => {
-    if (!data.length && (category || profile || search || id)) {
+    if (!data.length && (category || profile || search || id || recall)) {
       const url = {
         search: `/api/search?q=${q}&sort=${selected}`,
         profile: `/api/history/${profile}?sort=${selected}`,
         menu: `/api/menu/${c}?sort=${selected}`,
-      }[search ? "search" : profile ? "profile" : "menu"];
+        recall: `/api/threads?threads=${JSON.stringify(
+          splitarray(history, 0, 24)
+        )}`,
+      }[mode];
       axios
         .get(url)
         .then((res) => {
@@ -81,7 +91,10 @@ function MainContent() {
       search: `/api/search?q=${q}&sort=${selected}&page=${page + 1}`,
       profile: `/api/history/${profile}?sort=${selected}&page=${page + 1}`,
       menu: `/api/menu/${c}?sort=${selected}&page=${page + 1}`,
-    }[search ? "search" : profile ? "profile" : "menu"];
+      recall: `/api/threads?threads=${JSON.stringify(
+        splitarray(history, page * 25, (page + 1) * 25 - 1)
+      )}`,
+    }[mode];
     axios
       .get(url)
       .then((res) => {
@@ -118,16 +131,13 @@ function MainContent() {
     <Paper
       className="nobgimage noshadow overflow-auto"
       style={{
-        maxHeight: search ? "calc(100vh - 151px)" : "calc(100vh - 91px)",
+        maxHeight: `calc(100vh - ${
+          (search && "151") || (recall && "51") || "91"
+        }px)`,
       }}
       onScroll={onScroll}
     >
-      <Box
-        className="min-height-full"
-        sx={{
-          bgcolor: "primary.main",
-        }}
-      >
+      <Box className="min-height-full menu-bottom flex flex-dir-column">
         {!!(data.length && data?.[0] !== null) && (
           <Box className="flex flex-dir-column max-width-full menu-bottom">
             {data.map((thread: summary) => (
@@ -135,22 +145,28 @@ function MainContent() {
                 <MenuThread
                   key={`${category}${id === thread.id}`}
                   thread={thread}
+                  onClick={() => {
+                    const index = history.findIndex((i) => i === thread.id);
+                    if (index !== -1) history.splice(index, 1);
+                    history.unshift(thread.id);
+                    localStorage.setItem("history", JSON.stringify(history));
+                  }}
                 />
                 <Divider />
               </div>
             ))}
             {updating && <MenuPreload />}
-            {end && (
-              <Typography
-                className="mt10 mb10 text-align-center font-size-20-force"
-                sx={{
-                  color: "secondary.main",
-                }}
-              >
-                End
-              </Typography>
-            )}
           </Box>
+        )}
+        {end && (
+          <Typography
+            className="mt10 mb10 text-align-center font-size-20-force"
+            sx={{
+              color: "secondary.main",
+            }}
+          >
+            End
+          </Typography>
         )}
         {!data.length && <MenuPreload />}
       </Box>
@@ -165,6 +181,7 @@ function Menu() {
   const [query, setQuery] = useQuery();
   const [, setHistory] = useHistory();
   const [category] = useCat();
+  const [recall] = useRecall();
   const [profile] = useProfile();
   const navigate = useNavigate();
   const [n, setN] = useState(Math.random());
@@ -207,7 +224,7 @@ function Menu() {
           </div>
         </div>
       )}
-      <MainContent key={`${search}${profile}${category}${selected}${n}`} />
+      <MainContent key={`${search}${profile}${category}${recall}${selected}${n}`} />
     </Box>
   );
 }