+++ src/pages/create.tsx
@@ -11,6 +11,7 @@ import {
   SelectChangeEvent,
   TextField,
 } from "@mui/material";
+import { Create as CreateIcon } from "@mui/icons-material";
 import TextEditor from "../components/texteditor";
 import ReCAPTCHA from "react-google-recaptcha";
 import axios from "axios";
@@ -21,12 +22,15 @@ import {
   useData,
   useMenu,
   useProfile,
+  useRecall,
   useSearch,
+  useTitle,
 } from "../components/MenuProvider";
 import { useNotification, useWidth } from "../components/ContextProvider";
 import { categories, severity, wholepath } from "../lib/common";
 import MetahkgLogo from "../components/logo";
-/* A workaround for the recaptcha.reset() function not being exported from the recaptcha library. */
+import UploadImage from "../components/uploadimage";
+declare const tinymce: any;
 declare const grecaptcha: { reset: () => void };
 /**
  * It takes in a category number and a setter function for the category number, and returns a form
@@ -74,11 +78,14 @@ export default function Create() {
   const [profile, setProfile] = useProfile();
   const [cat, setCat] = useCat();
   const [search, setSearch] = useSearch();
+  const [recall, setRecall] = useRecall();
   const [data, setData] = useData();
+  const [mtitle, setMtitle] = useTitle();
   const [, setNotification] = useNotification();
-  const [catchoosed, setCatchoosed] = useState<number>(cat || 0);
+  const [catchoosed, setCatchoosed] = useState<number>(cat || 1);
   const [rtoken, setRtoken] = useState(""); //recaptcha token
   const [title, setTitle] = useState(""); //this will be the post title
+  const [imgurl, setImgurl] = useState("");
   const [icomment, setIcomment] = useState(""); //initial comment (#1)
   const [disabled, setDisabled] = useState(false);
   const [alert, setAlert] = useState<{ severity: severity; text: string }>({
@@ -95,24 +102,31 @@ export default function Create() {
    */
   useEffect(() => {
     if (localStorage.user && quote.id && quote.cid) {
+      setAlert({ severity: "info", text: "Fetching comment..." });
       setNotification({ open: true, text: "Fetching comment..." });
       axios
         .get(
           `/api/thread/${quote.id}?type=2&start=${quote.cid}&end=${quote.cid}`
         )
         .then((res) => {
-          if (res.data?.[0]) {
+          if (res.data?.[0]?.comment) {
             setInittext(
               `<blockquote style="color: #aca9a9; border-left: 2px solid #aca9a9; margin-left: 0"><div style="margin-left: 15px">${res.data?.[0]?.comment}</div></blockquote><p></p>`
             );
+            setAlert({ severity: "info", text: "" });
             setTimeout(() => {
               setNotification({ open: false, text: "" });
             }, 1000);
           } else {
+            setAlert({ severity: "error", text: "Comment not found!" });
             setNotification({ open: true, text: "Comment not found!" });
           }
         })
         .catch(() => {
+          setAlert({
+            severity: "warning",
+            text: "Unable to fetch comment. This comment would not be a quote.",
+          });
           setNotification({
             open: true,
             text: "Unable to fetch comment. This comment would not be a quote.",
@@ -123,6 +137,7 @@ export default function Create() {
   function create() {
     //send data to /api/create
     setAlert({ severity: "info", text: "Creating topic..." });
+    setNotification({ open: true, text: "Creating topic..." });
     setDisabled(true);
     axios
       .post("/api/create", {
@@ -134,9 +149,14 @@ export default function Create() {
       .then((res) => {
         cat && setCat(0);
         search && setSearch(false);
+        recall && setRecall(false);
         profile && setProfile(0);
         data.length && setData([]);
+        mtitle && setMtitle("");
         navigate(`/thread/${res.data.id}`);
+        setTimeout(() => {
+          setNotification({ open: false, text: "" });
+        }, 100);
       })
       .catch((err) => {
         setAlert({
@@ -164,7 +184,7 @@ export default function Create() {
   const small = width * 0.8 - 40 <= 450;
   return (
     <Box
-      className="flex fullwidth min-height-fullvh justify-center"
+      className="flex fullwidth min-height-fullvh justify-center max-width-full"
       sx={{
         backgroundColor: "primary.dark",
       }}
@@ -182,32 +202,74 @@ export default function Create() {
             <h1>Create topic</h1>
           </div>
           {alert.text && (
-            <Alert className="mt10 mb15" severity={alert.severity}>
+            <Alert className="mb15" severity={alert.severity}>
               {alert.text}
             </Alert>
           )}
-          <TextField
-            className="mb20"
-            variant="outlined"
-            color="secondary"
-            fullWidth
-            label="Title"
-            onChange={(e) => {
-              setTitle(e.target.value);
-            }}
-          />
+          <div className={width < 760 ? "" : "flex "}>
+            <ChooseCat cat={catchoosed} setCat={setCatchoosed} />
+            <TextField
+              className={width < 760 ? "mt15" : "ml15"}
+              variant="filled"
+              color="secondary"
+              fullWidth
+              label="Title"
+              onChange={(e) => {
+                setTitle(e.target.value);
+              }}
+            />
+          </div>
+          <div className="flex align-center mb15 mt15">
+            <UploadImage
+              onUpload={() => {
+                setAlert({ severity: "info", text: "Uploading image..." });
+                setNotification({ open: true, text: "Uploading image..." });
+              }}
+              onSuccess={(res) => {
+                setAlert({ severity: "info", text: "Image uploaded!" });
+                setNotification({ open: true, text: "Image uploaded!" });
+                setTimeout(() => {
+                  setNotification({ open: false, text: "" });
+                }, 1000);
+                setImgurl(res.data.url);
+                tinymce.activeEditor.insertContent(
+                  `<img src="${res.data.url}" style="object-fit: contain; max-width: 60%; max-height: 250px;" />`
+                );
+              }}
+              onError={() => {
+                setAlert({ severity: "error", text: "Error uploading image." });
+                setNotification({ open: true, text: "Error uploading image." });
+              }}
+            />
+            {imgurl && (
+              <p className="ml10 novmargin flex">
+                <a href={imgurl} target="_blank" rel="noreferrer">
+                  {imgurl}
+                </a>
+                <p
+                  className="link novmargin metahkg-grey-force ml5"
+                  onClick={() => {
+                    navigator.clipboard.writeText(imgurl);
+                    setNotification({
+                      open: true,
+                      text: "Copied to clipboard!",
+                    });
+                  }}
+                >
+                  copy
+                </p>
+              </p>
+            )}
+          </div>
           <TextEditor
             changehandler={(v, e: any) => {
               setIcomment(e.getContent());
             }}
             text={inittext}
           />
-          <div className="mt20">
-            <ChooseCat cat={catchoosed} setCat={setCatchoosed} />
-          </div>
           <div
-            className={`mt20 ${
-              small ? "" : "flex fullwidth justify-space-between"
+            className={`mt15 ${
+              small ? "" : "flex fullwidth justify-space-between align-center"
             }`}
           >
             <ReCAPTCHA
@@ -224,11 +286,14 @@ export default function Create() {
               disabled={
                 disabled || !(icomment && title && rtoken && catchoosed)
               }
-              className="mt20 font-size-16 create-btn"
+              className={`${
+                small ? "mt15 " : ""
+              }font-size-16-force create-btn novpadding notexttransform`}
               onClick={create}
               variant="contained"
               color="secondary"
             >
+              <CreateIcon className="mr5 font-size-16-force" />
               Create
             </Button>
           </div>