+++ src/pages/history.tsx
@@ -7,6 +7,9 @@ import {
   useProfile,
   useSearch,
   useSelected,
+  useTitle,
+  useData,
+  useRecall,
 } from "../components/MenuProvider";
 import { useHistory, useWidth } from "../components/ContextProvider";
 /**
@@ -18,21 +21,31 @@ export default function History() {
   const params = useParams();
   const [profile, setProfile] = useProfile();
   const [search, setSearch] = useSearch();
+  const [recall, setRecall] = useRecall();
   const [menu, setMenu] = useMenu();
   const [history, setHistory] = useHistory();
   const [width] = useWidth();
   const [selected, setSelected] = useSelected();
+  const [, setTitle] = useTitle();
+  const [, setData] = useData();
   const [id, setId] = useId();
   const [cat, setCat] = useCat();
   if (!(width < 760)) {
     return <Navigate to={`/profile/${params.id}`} replace />;
   }
+  function cleardata() {
+    setData([]);
+    setTitle("");
+    selected && setSelected(0);
+  }
   !menu && setMenu(true);
   history !== window.location.pathname && setHistory(window.location.pathname);
-  (profile !== Number(params.id) || "self") &&
+  (profile !== (Number(params.id) || "self") || search) &&
+    cleardata();
+  profile !== (Number(params.id) || "self") &&
     setProfile(Number(params.id) || "self");
   search && setSearch(false);
-  selected && setSelected(0);
+  recall && setRecall(false);
   id && setId(0);
   cat && setCat(0);
   return <div />;