+++ CHANGELOG.md
@@ -70,3 +70,16 @@ node server/migrate/migratetov0.5.2.js
 
 - story mode
 - create new topic from comment
+
+## v0.5.6rc1
+
+- image uploads using na.cx
+
+## v0.5.6rc2
+
+- support account verification in one week
+- support resend verification email
+
+## v0.5.6rc3
+
+- recall